Move symbols schema init specifics to SymbolsPartial
